Move Command classes to Command package
