+++ src/main/java/command/AddCommand.java
@@ -1,4 +1,6 @@
-package duke;
+package Command;
+
+import duke.*;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
