+++ src/main/java/command/InvalidCommandException.java
@@ -1,0 +1,9 @@
+package Command;
+
+import duke.DukeException;
+
+public class InvalidCommandException extends DukeException {
+    public InvalidCommandException(String msg) {
+        super(msg);
+    }
+}
