+++ src/main/java/command/MarkCommand.java
@@ -1,18 +1,20 @@
-package duke;
+package Command;
+
+import duke.*;
 
 import java.util.ArrayList;
 
-public class DeleteCommand extends Command {
-    public static final String COMMAND_DELETE = "delete";
-    public DeleteCommand(ArrayList<String> params) {
+public class MarkCommand extends Command {
+    public static final String COMMAND_MARK = "mark";
+    public MarkCommand(ArrayList<String> params) {
         super(params);
     }
 
     @Override
     public void execute(TaskList tasks, UI ui, Storage storage) throws DukeException {
         int index = Integer.parseInt(params.get(1)) - 1;
-        Task task = tasks.remove(index);
-        ui.printTaskDeletedMessage(task, tasks.getTaskCount());
+        Task task = tasks.mark(index);
+        ui.printTaskMarkedMessage(task);
         tasks.saveState(storage);
     }
