+++ src/main/java/duke/Duke.java
@@ -1,11 +1,9 @@
 package duke;
 
+import Command.Command;
+
 import java.io.FileNotFoundException;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.Scanner;
 
 public class Duke {
     private TaskList tasks;
