--- src/main/java/duke/InvalidCommandException.java
@@ -1,7 +1,0 @@
-package duke;
-
-public class InvalidCommandException extends DukeException{
-    public InvalidCommandException(String msg) {
-        super(msg);
-    }
-}
