+++ src/main/java/duke/Parser.java
@@ -1,6 +1,16 @@
 package duke;
 
-import java.io.IOException;
+import Command.AddCommand;
+import Command.Command;
+import Command.DeleteCommand;
+import Command.ExitCommand;
+import Command.FindCommand;
+import Command.InvalidCommandException;
+import Command.ListCommand;
+import Command.MarkCommand;
+import Command.OnCommand;
+import Command.UnmarkCommand;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
