+++ src/test/java/duke/TaskListTest.java
@@ -1,5 +1,6 @@
 package duke;
 
+import Command.InvalidCommandException;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
